cambio menor al tipo de pid asi se puede manejar el error con numeros negativos

+++ Userland/SampleCodeModule/shell.c
@@ -224,7 +224,7 @@ int main() {
 				int processIndex = id - builtinsDim;
 				int16_t fileDescriptors[] = {0, 1, 2};
 				uint64_t rip = (uint64_t)processCommands[processIndex].exec;
-				int pid = syscreateProcess(rip, (char **)params, processCommands[processIndex].name, 1, fileDescriptors);
+				int16_t pid = syscreateProcess(rip, (char **)params, processCommands[processIndex].name, 1, fileDescriptors);
 				syswaitProcess(pid);
 			} else {
 				printf(command);
